+++ spec/param_type_spec.rb
@@ -27,4 +27,26 @@ describe 'Params Types' do
       { 'paramType' => 'query', 'name' => 'input', 'description' => nil, 'type' => 'string', 'required' => true, 'allowMultiple' => false }
     ]
   end
+
+  describe 'header params' do
+    def app
+      Class.new(Grape::API) do
+        format :json
+
+        desc 'Some API', headers: { 'My-Header' => { required: true, description: 'Set this!' } }
+        params do
+          requires :input, type: String
+        end
+        post :action do
+        end
+
+        add_swagger_documentation
+      end
+    end
+
+    it 'has consistent types' do
+      types = subject.map { |param| param['type'] }
+      expect(types).to eq(%w(string string))
+    end
+  end
 end